# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandAccepted": {
        "properties": {
          "command_id": {
            "title": "Command Id",
            "type": "string"
          }
        },
        "required": [
          "command_id"
        ],
        "title": "CommandAccepted",
        "type": "object"
      },
      "CommandIn": {
        "properties": {
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "CommandIn",
        "type": "object"
      },
      "CommandOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "payload",
          "result",
          "requested_by",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "CommandOut",
        "type": "object"
      },
      "ConfigIn": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "yaml": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "yaml"
        ],
        "title": "ConfigIn",
        "type": "object"
      },
      "ConfigOut": {
        "properties": {
          "json_schema": {
            "additionalProperties": true,
            "title": "Json Schema",
            "type": "object"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          },
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "yaml",
          "sha256",
          "version_id",
          "json_schema"
        ],
        "title": "ConfigOut",
        "type": "object"
      },
      "ConfigSaved": {
        "properties": {
          "changed": {
            "title": "Changed",
            "type": "boolean"
          },
          "command_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Id"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "version_id",
          "changed",
          "command_id"
        ],
        "title": "ConfigSaved",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "aifund",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/commands/{command_id}": {
      "get": {
        "operationId": "command_api_commands__command_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "command_id",
            "required": true,
            "schema": {
              "title": "Command Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command",
        "tags": [
          "commands"
        ]
      }
    },
    "/api/config": {
      "get": {
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "commands"
        ]
      },
      "put": {
        "operationId": "put_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FieldError"
                  },
                  "title": "Response 422 Put Config Api Config Put",
                  "type": "array"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Put Config",
        "tags": [
          "commands"
        ]
      }
    },
...
    "/api/engine/commands": {
      "post": {
        "operationId": "engine_command_api_engine_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Engine Command",
        "tags": [
          "commands"
        ]
      }
    },
...
    "/api/positions/{position_id}/close": {
      "post": {
        "operationId": "close_position_api_positions__position_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "title": "Position Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Position",
        "tags": [
          "commands"
        ]
      }
    },
...
  }
}